fix checkmate and draw endings and change move logic to work with json object instead of string
