+++ src/frontend/package.json
@@ -9,7 +9,8 @@
   "dependencies": {
     "@chrisoakman/chessboardjs": "^1.0.0",
     "chess.js": "v1.0.0-beta.8",
-    "requirejs": "^2.3.7"
+    "requirejs": "^2.3.7",
+    "stockfish.wasm": "^0.10.0"
   },
   "keywords": [],
   "author": "",
