+++ src/frontend/static/css/game.css
@@ -1283,7 +1283,6 @@ button:focus {
   display: none;
   background-color: #17477E;
   background-size: cover;
-  #background-image: url("../img/wood-pattern2.png");
   position: absolute;
   width: 100%;
   height: 100%;
@@ -1562,7 +1561,7 @@ th, td {
 
 .whiteCell {
   background-color: white;
-  width: 5.25%;
+  width: 5%;
 
 }
 .blackCell {
@@ -1572,6 +1571,27 @@ th, td {
 .moveTableHeader {
   width: 100%;
 }
+
+/* width */
+.tableWrapper::-webkit-scrollbar {
+  width: 15px;
+}
+
+/* Track */
+.tableWrapper::-webkit-scrollbar-track {
+  background: white;
+}
+
+/* Handle */
+.tableWrapper::-webkit-scrollbar-thumb {
+  background: #E1ECE0;
+}
+
+/* Handle on hover */
+.tableWrapper::-webkit-scrollbar-thumb:hover {
+  background: #999;
+}
+
 .moveTable {
   width: 99%;
   position: relative;
@@ -1585,33 +1605,6 @@ th, td {
   float: left;
   display: none;
 }
-
-/*.myButton {
-    width: 100px;
-    margin: 15px;
-    padding: 10px;
-    font-size: 14px;
-    color: white;
-    text-align: center;
-    font-family: "Baloo Paaji 2";
-    #float: right;
-    border-radius: 20px;
-    background: #fc5185;
-    #animation: glowingButton 1.5s infinite;
-}
-
-.myButton:hover {
-      color: white;
-      cursor: pointer;
-      #background-image: linear-gradient(to right, #fc5185 0%, #FCBBCD 100%), radial-gradient(circle at top left, #fc5185, #FCBBCD);
-      #background-image: -webkit-linear-gradient(to right, #fc5185 0%, #FCBBCD 100%), -webkit-radial-gradient(circle at top left, #fc5185, #FCBBCD);
-      #background-image: -moz-linear-gradient(to right, #fc5185 0%, #FCBBCD 100%), -moz-radial-gradient(circle at top left, #fc5185, #FCBBCD);
-      #background-image: -ms-linear-gradient(to right, #fc5185 0%, #FCBBCD 100%), -ms-radial-gradient(circle at top left, #fc5185, #FCBBCD);
-      #background-image: -o-linear-gradient(to right, #fc5185 0%, #FCBBCD 100%), -o-radial-gradient(circle at top left, #fc5185, #FCBBCD);
-      #background-image: -o-linear-gradient(to right, #fc5185 0%, #FCBBCD 100%), -o-radial-gradient(circle at top left, #fc5185, #FCBBCD);
-      color: #fc5185;
-      background: white;
-}*/
 
 .myButtonClone {
     width: 100px;
