+++ src/frontend/static/html/game.html
@@ -14,7 +14,6 @@
         <link rel="stylesheet" href="../css/magnific-popup.css">
         <link rel="stylesheet" href="../vendors/nice-select/css/nice-select.css">
         <link rel="stylesheet" href="../vendors/animate-css/animate.css">
-        <link rel="stylesheet" href="../vendors/flaticon/flaticon.css">
         <!-- main css -->
         <link rel="stylesheet" href="../css/style.css">
         <link rel="stylesheet" href="../css/game.css">
@@ -252,7 +251,6 @@
         </script>
         <script src="../vendors/counter-up/jquery.waypoints.min.js"></script>
         <script src="../vendors/counter-up/jquery.counterup.min.js"></script>
-        <script src="../js/mail-script.js"></script>
         <script src="../vendors/lightbox/simpleLightbox.min.js"></script>
         <!-- contact js -->
         <script>var exports = {};</script>
