+++ src/frontend/static/html/pricing.html
@@ -8,15 +8,10 @@
         <title>Chessmine</title>
         <!-- Bootstrap CSS -->
         <link rel="stylesheet" href="../css/bootstrap.css">
-        <link rel="stylesheet" href="../vendors/linericon/style.css">
         <link rel="stylesheet" href="../css/font-awesome.min.css">
-        <link rel="stylesheet" href="../vendors/owl-carousel/owl.carousel.min.css">
-        <link rel="stylesheet" href="../vendors/lightbox/simpleLightbox.css">
-        <link rel="stylesheet" href="../vendors/nice-select/css/nice-select.css">
         <link rel="stylesheet" href="../vendors/animate-css/animate.css">
         <!-- main css -->
         <link rel="stylesheet" href="../css/style.css">
-        <link rel="stylesheet" href="static/css/responsive.css">
     </head>
     <body>
 	<!--================Header Menu Area =================-->
@@ -107,20 +102,14 @@
             crossorigin="anonymous"
             referrerpolicy="no-referrer">
     </script>
-    <script src="../vendors/lightbox/simpleLightbox.min.js"></script>
-    <script src="../vendors/nice-select/js/jquery.nice-select.min.js"></script>
-    <script src="../vendors/isotope/imagesloaded.pkgd.min.js"></script>
-    <script src="../vendors/isotope/isotope-min.js"></script>
-    <script src="../vendors/owl-carousel/owl.carousel.min.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-ajaxchimp/1.3.0/jquery.ajaxchimp.min.js"
             integrity="sha512-5yj5elY9u6clGe9/97bj3jJlw8+O9XSv/tbme8m/LR8cKnnT5+rR8qHW/UYQ/MozLg3cvTHeYIpM5kRktASSbg=="
             crossorigin="anonymous"
             referrerpolicy="no-referrer">
     </script>
-    <script src="../js/mail-script.js"></script>
     <script src="../vendors/counter-up/jquery.waypoints.min.js"></script>
     <script src="../vendors/counter-up/jquery.counterup.min.js"></script>
-    <script src="js/theme.js"></script>
+    <script src="../js/theme.js"></script>
     <script type="text/javascript">
             $(function(){
               $("#header").load("html/header.html");
