+++ src/frontend/static/html/settings.html
@@ -8,17 +8,12 @@
         <title>Chessmine</title>
         <!-- Bootstrap CSS -->
         <link rel="stylesheet" href="../css/bootstrap.css">
-        <link rel="stylesheet" href="../vendors/linericon/style.css">
         <link rel="stylesheet" href="../css/font-awesome.min.css">
-        <link rel="stylesheet" href="../vendors/owl-carousel/owl.carousel.min.css">
         <link rel="stylesheet" href="../css/magnific-popup.css">
-        <link rel="stylesheet" href="../vendors/nice-select/css/nice-select.css">
         <link rel="stylesheet" href="../vendors/animate-css/animate.css">
-        <link rel="stylesheet" href="../vendors/flaticon/flaticon.css">
         <!-- main css -->
         <link rel="stylesheet" href="../css/style.css">
         <link rel="stylesheet" href="../css/settings.css">
-        <link rel="stylesheet" href="../css/responsive.css">
         <link rel="stylesheet" href="../css/loading.css">
         <link href='https://fonts.googleapis.com/css?family=Orbitron' rel='stylesheet' type='text/css'>
         <!--link rel="stylesheet" href="static/css/clock.css"-->
@@ -238,10 +233,6 @@
                 referrerpolicy="no-referrer">
         </script>
         <script src="../js/theme.js"></script>
-        <script src="../vendors/nice-select/js/jquery.nice-select.min.js"></script>
-        <script src="../vendors/isotope/imagesloaded.pkgd.min.js"></script>
-        <script src="../vendors/isotope/isotope-min.js"></script>
-        <script src="../vendors/owl-carousel/owl.carousel.min.js"></script>
         <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-ajaxchimp/1.3.0/jquery.ajaxchimp.min.js"
                 integrity="sha512-5yj5elY9u6clGe9/97bj3jJlw8+O9XSv/tbme8m/LR8cKnnT5+rR8qHW/UYQ/MozLg3cvTHeYIpM5kRktASSbg=="
                 crossorigin="anonymous"
@@ -249,8 +240,6 @@
         </script>
         <script src="../vendors/counter-up/jquery.waypoints.min.js"></script>
         <script src="../vendors/counter-up/jquery.counterup.min.js"></script>
-        <script src="../js/mail-script.js"></script>
-        <script src="../vendors/lightbox/simpleLightbox.min.js"></script>
         <!-- contact js -->
         <script>var exports = {};</script>
         <script src="../js/chess.min.js"></script>
