--- src/frontend/static/vendors/flaticon/_flaticon.scss
@@ -1,45 +1,0 @@
-    /*
-    Flaticon icon font: Flaticon
-    Creation date: 26/06/2018 17:19
-    */
-
-    @font-face {
-  font-family: "Flaticon";
-  src: url("./Flaticon.eot");
-  src: url("./Flaticon.eot?#iefix") format("embedded-opentype"),
-       url("./Flaticon.woff") format("woff"),
-       url("./Flaticon.ttf") format("truetype"),
-       url("./Flaticon.svg#Flaticon") format("svg");
-  font-weight: normal;
-  font-style: normal;
-}
-
-@media screen and (-webkit-min-device-pixel-ratio:0) {
-  @font-face {
-    font-family: "Flaticon";
-    src: url("./Flaticon.svg#Flaticon") format("svg");
-  }
-}
-
-    .fi:before{
-        display: inline-block;
-  font-family: "Flaticon";
-  font-style: normal;
-  font-weight: normal;
-  font-variant: normal;
-  line-height: 1;
-  text-decoration: inherit;
-  text-rendering: optimizeLegibility;
-  text-transform: none;
-  -moz-osx-font-smoothing: grayscale;
-  -webkit-font-smoothing: antialiased;
-  font-smoothing: antialiased;
-    }
-
-    .flaticon-skyline:before { content: "\f100"; }
-.flaticon-sketch:before { content: "\f101"; }
-.flaticon-city:before { content: "\f102"; }
-    
-    $font-Flaticon-skyline: "\f100";
-    $font-Flaticon-sketch: "\f101";
-    $font-Flaticon-city: "\f102";
